Register `SideNote` and `Highlight` as separate components
MDX 2 started “sandboxing” elements. So using `<aside>` as a tag in the content isn't now replaced with the `SideNote` component because it was used as a “component”.

@see https://mdxjs.com/migrating/v2/#update-mdx-content

Markdown doesn't have a native way to represent `aside` and `mark` so, without a plugin, I will use components in the content for now.

Later I might change the content back to be more framework-agnostic.