Specify adapter types for the module
Again, these aren't “real” adapters. It would be better and cleaner
to create an ACL where the outside world interfaces
would be adapted to the application needs.

However, I don't think it's that important in a pet project.

I can allow myself sacrifice ”purity“ in favor of brevity and speed.

I will consider writing these in a correct way if I ever need to replace
the underlying thrird-party service.